refactor: extract scrollspy in own component
