+++ src/pages/MarkingPage/index.tsx
@@ -1,15 +1,14 @@
 import { ChevronUpIcon } from '@radix-ui/react-icons'
 import { Box, Card, Flex, Grid, Section as RadixUISection, Text } from '@radix-ui/themes'
-import { isNil, keyBy, mapValues, sumBy, values } from 'lodash'
+import { keyBy, mapValues, sumBy, values } from 'lodash'
 import React, { FC, useMemo, useState } from 'react'
 
 import Body from '../../components/pageStructure/Body'
 import MarkingToolbar from '../../components/topBars/MarkingToolbar'
 import { useAnswers, useMarks, useQuestions, useStudents } from '../../hooks/marking'
 import { Student } from '../../types/marking'
-import { numberToLetter, numberToRoman } from '../../utils/common'
 import MarkableSubmission from './MarkableSubmission'
-import { ScrollspyItem } from './ScrollspyItem'
+import Scrollspy from './Scrollspy'
 import './index.css'
 
 const MarkingPage: FC = () => {
@@ -67,40 +66,12 @@ const MarkingPage: FC = () => {
               />
             )}
           </Box>
-          <Flex direction="column" gap="8" p="4" className="sticky-sidebar">
-            {student &&
-              Object.entries(questions).map(([q_, question]) => {
-                const q = Number(q_)
-                const partial = sumBy(
-                  rawMarksTable[student.username].filter((m) => m.question === q && !isNil(m.mark)),
-                  'mark'
-                )
 
-                return (
-                  <Flex direction="column" gap="2">
-                    <ScrollspyItem
-                      label={`Question ${q}`}
-                      total={question.availableMarks}
-                      partial={partial}
-                    />
-                    {Object.entries(question.parts).map(([p_, part]) => {
-                      const p = Number(p_)
-                      return Object.entries(part.sections).map(([s_, section]) => {
-                        const s = Number(s_)
-                        return (
-                          <ScrollspyItem
-                            label={`Part ${numberToLetter(p)} ${numberToRoman(s)}`}
-                            total={section.maximumMark}
-                            partial={lookupMark(student.username, q, p, s)?.mark}
-                            indent={true}
-                          />
-                        )
-                      })
-                    })}
-                  </Flex>
-                )
-              })}
-          </Flex>
+          {student && (
+            <Box className="sticky-sidebar">
+              <Scrollspy questions={questions} marks={rawMarksTable[student.username]} />
+            </Box>
+          )}
         </Grid>
       </RadixUISection>
     </>
